Move recipient validation to domain layer
