+++ src/blueferry/obex/map_send.py
@@ -16,7 +16,6 @@ from __future__ import annotations
 
 import logging
 import os
-import re
 from collections.abc import Sequence
 
 import dbus
@@ -26,65 +25,9 @@ from blueferry.bus import obex
 from blueferry.limits import MAX_OUTGOING_BODY_BYTES
 from blueferry.obex.transfer import wait_for_transfer
 from blueferry.private_files import create_runtime_private_file
+from blueferry.recipients import InvalidRecipient, validate_recipient
 
 log = logging.getLogger(__name__)
-
-# A recipient is interpolated straight into a bMessage vCard property, so it
-# must not be able to carry bMessage structure. Keep both accepted forms
-# deliberately narrower than their full formal grammars: this is an address
-# for Messages, not a general-purpose vCard parser.
-_PHONE_RE = re.compile(r"^\+?[0-9][0-9 ()\-.]{2,29}$")
-_EMAIL_LOCAL_RE = re.compile(r"^[A-Za-z0-9!#$%&'*+/=?^_`{|}~.-]+$")
-_EMAIL_DOMAIN_LABEL_RE = re.compile(r"^[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?$")
-
-
-class InvalidRecipient(ValueError):
-    """Raised when a recipient can't be safely placed in a bMessage."""
-
-
-def _normalize_email(candidate: str) -> str | None:
-    """Return a conservative ASCII email address, or None if invalid."""
-    if len(candidate) > 254 or candidate.count("@") != 1:
-        return None
-    local, domain = candidate.rsplit("@", 1)
-    if not local or len(local) > 64 or not domain or len(domain) > 253:
-        return None
-    if (local.startswith(".") or local.endswith(".") or ".." in local
-            or not _EMAIL_LOCAL_RE.fullmatch(local)):
-        return None
-    labels = domain.split(".")
-    if any(not _EMAIL_DOMAIN_LABEL_RE.fullmatch(label) for label in labels):
-        return None
-    # DNS names are case-insensitive. Preserve the local part because it is
-    # technically allowed to be case-sensitive.
-    return f"{local}@{domain.lower()}"
-
-
-def validate_recipient(recipient: str) -> str:
-    """Return a normalized phone number or email address.
-
-    Without this, a recipient containing CRLF injects arbitrary bMessage
-    structure — e.g. a second recipient vCard, so the message silently goes
-    to an extra address. Reject rather than attempting to escape an invalid
-    recipient.
-    """
-    candidate = (recipient or "").strip()
-    if _PHONE_RE.fullmatch(candidate):
-        # Collapse to `+`-and-digits so the wire format is predictable.
-        digits = re.sub(r"[^0-9]", "", candidate)
-        if not 3 <= len(digits) <= 20:
-            raise InvalidRecipient(
-                f"implausible phone-number length: {candidate!r}")
-        return ("+" if candidate.startswith("+") else "") + digits
-
-    email = _normalize_email(candidate)
-    if email is not None:
-        return email
-
-    raise InvalidRecipient(
-        f"not a valid phone number or email address: {candidate!r}"
-    )
-
 
 def _byte_stuff(body: str) -> str:
     """Apply MAP bMessage byte-stuffing.
